feat(rdkronos): Introduce FW_CONFIG loading for BL2_AT_EL3
The FW_CONFIG is typically loaded by BL1 and passed as an argument to
BL2. When using BL2_AT_EL3, there is no BL1 so it is assumed that all
configuration is provided at compile time.

For the rdkronos platform, introduce the ability to load the FW_CONFIG
from the FIP as part of BL2 (at EL3) by overriding
bl2_el3_plat_arch_setup.

Upstream-Status: Pending
Signed-off-by: Peter Hoyes <Peter.Hoyes@arm.com>