feat(rdkronos): Enable SVE_FOR_NS and SVE_FOR_SWD flags
Enable support for SVE on RD-Kronos platforms.

SVE has been built into the Kronos FVP, but the TF-A disable the SVE
support for Secure and Non-Secure world by default. This will cause
the SVE, SMID and Float-Point access code in S-EL1/EL0 enter unexcepted
trap.

    Unhandled Exception in EL3
    x30 = 0x0000000000011888
    x0 = 0x0000000000300000
    x1 = 0x0000000000000001
    ...

This patch enables the SVE support for Kronos by default in TF-A.

Upstream-Status: Pending
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Robbie Cao <robbie.cao@arm.com>